Reverted latest modifications.
The updates to the slides looked terrible and did not add any valuable information. Nobody wants to read source code like this, neither during a real talk nor on the screen. You can open the examples in a parallel editor.

The HTML version was even more unusable.

I will add a PDF version to the website.

Sorry, eh..


git-svn-id: http://daimlerstr.de/svn/lab/trunk@621 d7904494-db04-0410-88ca-e47d1dcc2d9a